Добавлено исправление для некорректных значений datetime в базе данных MySQL. Обновлены модели с правильными GORM тегами для автоматического управления временными метками. Реализованы функции для исправления существующих некорректных значений при миграции. Добавлены новые миграционные скрипты и тесты для проверки исправлений. Обновлены зависимости в go.mod. Улучшена логика фильтрации моделей для отображения только включенных по умолчанию.

+++ backend/internal/litellm/client.go
@@ -41,8 +41,8 @@ type LiteLLMModelGroupResponse struct {
 type LiteLLMModelGroup struct {
 	ModelGroup                       string   `json:"model_group"`
 	Providers                        []string `json:"providers"`
-	MaxInputTokens                   int      `json:"max_input_tokens"`
-	MaxOutputTokens                  int      `json:"max_output_tokens"`
+	MaxInputTokens                   float64  `json:"max_input_tokens"`
+	MaxOutputTokens                  float64  `json:"max_output_tokens"`
 	InputCostPerToken                float64  `json:"input_cost_per_token"`
 	OutputCostPerToken               float64  `json:"output_cost_per_token"`
 	Mode                             string   `json:"mode"`
